Renamed private to _private; updated the armc templates
