--- FOX/data/armc_template.yaml
@@ -1,41 +1,0 @@
-param:
-    prm_file: null
-
-psf:
-    str_file: null
-    rtf_file: null
-    psf_file: null
-    ligand_atoms: null
-
-pes: null
-
-molecule: null
-
-job:
-    logfile: armc.log
-    job_type: scm.plams.Cp2kJob
-    name: armc
-    path: .
-    folder: MM_MD_workdir
-    keep_files: False
-    rmsd_threshold: 10.0
-    preopt_settings: null
-    md_settings: null
-
-hdf5_file: armc.hdf5
-
-armc:
-    iter_len: 50000
-    sub_iter_len: 100
-    gamma: 2.0
-    a_target: 0.25
-    phi: 1.0
-
-move:
-    func: numpy.multiply
-    kwargs: {}
-    args: []
-    range:
-        start: 0.005
-        stop: 0.1
-        step: 0.005
