--- FOX/data/md_cp2k_template.yaml
@@ -1,40 +1,0 @@
-input:
-    global:
-        run_type: MD
-
-    force_eval:
-        method: FIST
-        mm:
-            forcefield:
-                ei_scale14: 1.0
-                vdw_scale14: 1.0
-                ignore_missing_critical_params: ''
-                parmtype: CHM
-                do_nonbonded: ''
-                shift_cutoff: .TRUE.
-                spline:
-                    emax_spline: 10e10
-            poisson:
-                periodic: NONE
-                ewald:
-                    ewald_type: NONE
-        subsys:
-            cell:
-                abc: '[angstrom] 50.0 50.0 50.0'
-                periodic: NONE
-            topology:
-                coord_file_format: 'OFF'
-                center_coordinates:
-                    center_point: 0.0 0.0 0.0
-
-    motion:
-        md:
-            ensemble: NVT
-            temperature: 300.0
-            timestep: 1.0
-            time_start_val: 5000
-            steps: 10000
-            thermostat:
-                type: CSVR
-                csvr:
-                    timecon: 1250
