+++ FOX/examples/_private/armc.yaml
