+++ FOX/examples/_private/armc_ivan.yaml
