+++ FOX/examples/_private/armc_ivan_bulk.yaml
