+++ FOX/examples/_private/armc_new.yaml
@@ -1,7 +1,7 @@
 param:
     type: FOX.armc.ParamMapping
     move_range:
-        ratio: null
+        ratio: [1, 5]
         start: 0.005
         stop: 0.1
         step: 0.005
@@ -11,12 +11,12 @@ param:
     charge:
         param: charge
         constraints:
-            - 0.5 < Cd < 1.5
-            - -0.5 > Se > -1.5
-            - 0 > O_1 > -1
+            - '0.5 < Cd < 1.5'
+            - '-0.5 > Se > -1.5'
+            - '0 > O_1 > -1'
         Cd: 0.9768
         Se: -0.9768
-        O_1: -0.4704
+        O_1: -0.47041
         frozen:
             C_1: 0.4524
     lennard_jones:
@@ -49,35 +49,36 @@ job:
     type: FOX.armc.PackageManager
     molecule: /Users/bvanbeek/Documents/GitHub/auto-FOX/FOX/data/Cd68Se55_26COO_MD_trajec.xyz
 
-    geometry:
-        type: qmflows.cp2k_mm
-        template: qmflows.geometry.specific.cp2k_mm
-        settings:
-            prm: ligand.prm
     md:
         type: qmflows.cp2k_mm
         template: qmflows.md.specific.cp2k_mm
         settings:
             prm: ligand.prm
             input:
+                force_eval:
+                    mm:
+                        forcefield:
+                            spline:
+                                emax_spline: 500.0
+                                r0_nb: 0.1
                 motion:
                     md:
                         steps: 100
 
 monte_carlo:
-    type: FOX.armc.ARMC
+    type: FOX.armc.ARMCPT
     iter_len: 50000
-    sub_iter_len: 100
+    sub_iter_len: 10
     logfile: armc.log
     hdf5_file: armc.hdf5
     path: .
     folder: MM_MD_workdir
-    keep_files: False
+    keep_files: True
 
 phi:
     type: FOX.armc.PhiUpdater
-    phi: 1.0
-    gamma: 2.0
-    a_target: 0.25
+    phi: [1, 5]
+    gamma: [2, 2]
+    a_target: [0.25, 0.5]
     func: numpy.add
     kwargs: {}
