+++ FOX/examples/_private/ligand.prm
