+++ FOX/examples/_private/ligand.str
