+++ FOX/examples/_private/test_armc.py
