+++ FOX/examples/armcpt.yaml
@@ -1,6 +1,7 @@
 param:
     type: FOX.armc.ParamMapping
     move_range:
+        ratio: [1, 5]
         start: 0.005
         stop: 0.1
         step: 0.005
@@ -55,7 +56,7 @@ job:
             prm: ligand.prm
 
 monte_carlo:
-    type: FOX.armc.ARMC
+    type: FOX.armc.ARMCPT
     iter_len: 50000
     sub_iter_len: 10
     logfile: armc.log
@@ -66,8 +67,8 @@ monte_carlo:
 
 phi:
     type: FOX.armc.PhiUpdater
-    phi: 1
-    gamma: 2
-    a_target: 0.25
+    phi: [1, 5]
+    gamma: [2, 2]
+    a_target: [0.25, 0.5]
     func: numpy.add
     kwargs: {}
